move pandas support into basic.py
