+++ python-package/lightgbm/basic.py
@@ -358,6 +358,39 @@ class Predictor(object):
             raise ValueError("incorrect number for predict result")
         return preds, nrow
 
+# pandas
+try:
+    from pandas import DataFrame
+except ImportError:
+    class DataFrame(object):
+        pass
+
+PANDAS_DTYPE_MAPPER = {'int8': 'int', 'int16': 'int', 'int32': 'int', 'int64': 'int',
+                       'uint8': 'int', 'uint16': 'int', 'uint32': 'int', 'uint64': 'int',
+                       'float16': 'float', 'float32': 'float', 'float64': 'float',
+                       'bool': 'i'}
+
+def _data_from_pandas(data):
+    if isinstance(data, DataFrame):
+        data_dtypes = data.dtypes
+        if not all(dtype.name in PANDAS_DTYPE_MAPPER for dtype in data_dtypes):
+            bad_fields = [data.columns[i] for i, dtype in
+                          enumerate(data_dtypes) if dtype.name not in PANDAS_DTYPE_MAPPER]
+
+            msg = """DataFrame.dtypes for data must be int, float or bool. Did not expect the data types in fields """
+            raise ValueError(msg + ', '.join(bad_fields))
+        data = data.values.astype('float')
+    return data
+
+def _label_from_pandas(label):
+    if isinstance(label, DataFrame):
+        if len(label.columns) > 1:
+            raise ValueError('DataFrame for label cannot have multiple columns')
+        label_dtypes = label.dtypes
+        if not all(dtype.name in PANDAS_DTYPE_MAPPER for dtype in label_dtypes):
+            raise ValueError('DataFrame.dtypes for label must be int, float or bool')
+        label = label.values.astype('float')
+    return label
 
 class Dataset(object):
     """Dataset used in LightGBM.
@@ -398,6 +431,8 @@ class Dataset(object):
         if data is None:
             self.handle = None
             return
+        data = _data_from_pandas(data)
+        label = _label_from_pandas(label)
         self.data_has_header = False
         """process for args"""
         params = {} if params is None else params
